+++ Makefile
@@ -26,10 +26,6 @@ lint: $(all)
 		$(MAKE) lint -C $$dir || exit 1; \
 	done
 
-test-integration:
-	@echo Running integration tests
-	@cd ./e2e && make
-
 test: $(all)
 	for dir in $(all); do \
 		$(MAKE) test -C $$dir || exit 1; \
