+++ core/pkg/decorators/keyvalue/keyvalue_suite_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package document_test
+package keyvalue_test
 
 import (
 	"testing"
