--- e2e/Makefile
@@ -1,5 +1,0 @@
-test-integration: test-integration-document
-
-test-integration-document:
-	@echo Running document integration tests
-	@go run github.com/onsi/ginkgo/ginkgo ./document/...
