--- e2e/document/set.go
@@ -1,102 +1,0 @@
-// Copyright 2021 Nitric Pty Ltd.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package document_suite
-
-import (
-	"context"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"github.com/nitrictech/nitric/core/pkg/plugins/document"
-)
-
-func SetTests(docPlugin document.DocumentService) {
-	Context("Set", func() {
-		When("Blank key.Collection.Name", func() {
-			It("Should return error", func() {
-				key := document.Key{Id: "1"}
-				err := docPlugin.Set(context.TODO(), &key, UserItem1)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Blank key.Id", func() {
-			It("Should return error", func() {
-				key := document.Key{Collection: &document.Collection{Name: "users"}}
-				err := docPlugin.Set(context.TODO(), &key, UserItem1)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Nil item map", func() {
-			It("Should return error", func() {
-				key := document.Key{Collection: &document.Collection{Name: "users"}, Id: "1"}
-				err := docPlugin.Set(context.TODO(), &key, nil)
-				Expect(err).Should(HaveOccurred())
-			})
-		})
-		When("Valid New Set", func() {
-			It("Should store new item successfully", func() {
-				err := docPlugin.Set(context.TODO(), &UserKey1, UserItem1)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(context.TODO(), &UserKey1)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content["email"]).To(BeEquivalentTo(UserItem1["email"]))
-			})
-		})
-		When("Valid Update Set", func() {
-			It("Should update existing item successfully", func() {
-				err := docPlugin.Set(context.TODO(), &UserKey1, UserItem1)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(context.TODO(), &UserKey1)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content["email"]).To(BeEquivalentTo(UserItem1["email"]))
-
-				err = docPlugin.Set(context.TODO(), &UserKey1, UserItem2)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err = docPlugin.Get(context.TODO(), &UserKey1)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content["email"]).To(BeEquivalentTo(UserItem2["email"]))
-			})
-		})
-		When("Valid Sub Collection Set", func() {
-			It("Should store item successfully", func() {
-				err := docPlugin.Set(context.TODO(), &Customer1.Orders[0].Key, Customer1.Orders[0].Content)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(context.TODO(), &Customer1.Orders[0].Key)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content).To(BeEquivalentTo(Customer1.Orders[0].Content))
-			})
-		})
-		When("Valid Multiple Sub Collection Set", func() {
-			It("Should store item successfully", func() {
-				err := docPlugin.Set(context.TODO(), &Customer1.Reviews[0].Key, Customer1.Reviews[0].Content)
-				Expect(err).ShouldNot(HaveOccurred())
-
-				doc, err := docPlugin.Get(context.TODO(), &Customer1.Reviews[0].Key)
-				Expect(err).ShouldNot(HaveOccurred())
-				Expect(doc).ToNot(BeNil())
-				Expect(doc.Content).To(BeEquivalentTo(Customer1.Reviews[0].Content))
-			})
-		})
-	})
-}
